DR_Webapp/requirements.txt
flask
torch
torchvision
pillow
werkzeug
